requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile .\requirements.in --output-file ./requirements.txt
pillow==11.3.0
    # via -r ./requirements.in
